Fix file name to match migrator pattern
